Add custom exceptions and contoller advisor

+++ src/main/java/com/ufar/studentmanagementsystem/service/impl/UserServiceImpl.java
@@ -1,8 +1,12 @@
 package com.ufar.studentmanagementsystem.service.impl;
 
+import com.ufar.studentmanagementsystem.exception.already_exists_exception.UserAlreadyExistsException;
+import com.ufar.studentmanagementsystem.exception.invalid_exception.InvalidUserException;
+import com.ufar.studentmanagementsystem.exception.not_found_exception.UserNotFoundException;
 import com.ufar.studentmanagementsystem.model.User;
 import com.ufar.studentmanagementsystem.repository.UserRepository;
 import com.ufar.studentmanagementsystem.service.UserService;
+import com.ufar.studentmanagementsystem.utils.validation.UserValidation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,6 +28,17 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public User addUser(User user) {
+        if (!UserValidation.IsValid(user)) {
+            //logging The user is not valid
+            throw new InvalidUserException("The user is not valid");
+        }
+
+        User existingUser = findUserByUsername(user.getUserName()).orElse(null);
+
+        if (existingUser != null) {
+            //logging "The user already exists with username " + user.getUserName()
+            throw new UserAlreadyExistsException("The user already exists with username " + user.getUserName());
+        }
         return userRepository.add(user);
     }
 
@@ -33,19 +48,50 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> findUserById(Integer id) {
-        return userRepository.findById(id);
+    public User findUserById(Integer id) {
+        User user = userRepository.findById(id).orElse(null);
+        if (user == null) {
+            //logging "The user with id " + id + " is not found"
+            throw new UserNotFoundException("The user with id " + id + " is not found");
+        }
+        return user;
+    }
+
+    @Override
+    public Optional<User> findUserByUsername(String username) {
+        User user = userRepository.findByUsername(username).orElse(null);
+        if (user == null) {
+            //logging "The user with username " + username + " is not found"
+            throw new UserNotFoundException("The user with username " + username + " is not found");
+        }
+        return Optional.of(user);
     }
 
     @Override
     @Transactional
-    public Optional<User> updateUser(User user) {
-        return userRepository.update(user);
+    public User updateUser(User user) {
+        if (!UserValidation.IsValid(user) || user.getId() <= 0 || user.getId() == null) {
+            //logging The user is not valid
+            throw new InvalidUserException("The user is not valid");
+        }
+
+        User updatingUser = userRepository.findById(user.getId()).orElse(null);
+
+        if (updatingUser == null) {
+            //logging "The user with id " + user.getId() + " is not found"
+            throw new UserNotFoundException("The user with id " + user.getId() + " is not found");
+        }
+        return userRepository.update(user).get();
     }
 
     @Override
     @Transactional
     public void deleteUserById(Integer id) {
+        User updatingUser = userRepository.findById(id).orElse(null);
+        if (updatingUser == null) {
+            //logging "The user with id " + id + " is not found"
+            throw new UserNotFoundException("The user with id " + id + " is not found");
+        }
         userRepository.deleteById(id);
     }
 }
